chore: move things around a little bit to the app plugin
