+++ plugins/app/src/extensions/AppLanguageApi.ts
@@ -15,7 +15,7 @@
  */
 
 // eslint-disable-next-line @backstage/no-relative-monorepo-imports
-import { AppLanguageSelector } from '../../../core-app-api/src/apis/implementations/AppLanguageApi';
+import { AppLanguageSelector } from '../../../../packages/core-app-api/src/apis/implementations/AppLanguageApi';
 import { appLanguageApiRef } from '@backstage/core-plugin-api/alpha';
 import { ApiBlueprint, createApiFactory } from '@backstage/frontend-plugin-api';
