+++ plugins/app/src/extensions/AppThemeApi.tsx
